Move paths helpers into their own file
