+++ packages/core/src/helpers/paths.ts
