+++ packages/plugins/build-report/src/rollup.ts
@@ -2,7 +2,7 @@
 // This product includes software developed at Datadog (https://www.datadoghq.com/).
 // Copyright 2019-Present Datadog, Inc.
 
-import { getAbsolutePath } from '@dd/core/helpers';
+import { getAbsolutePath } from '@dd/core/helpers/paths';
 import type { Logger, Entry, GlobalContext, Input, Output, PluginOptions } from '@dd/core/types';
 
 import { cleanName, cleanPath, cleanReport, getType } from './helpers';
