+++ packages/tests/src/_jest/helpers/mocks.ts
@@ -2,7 +2,7 @@
 // This product includes software developed at Datadog (https://www.datadoghq.com/).
 // Copyright 2019-Present Datadog, Inc.
 
-import { getAbsolutePath } from '@dd/core/helpers';
+import { getAbsolutePath } from '@dd/core/helpers/paths';
 import type {
     BuildReport,
     File,
